fix: resolve orientation, scale, and depth processing issues
Testing & Validation:
- Added `--calibrate` / `-c` flag to `scripts/zmq_rotate_test.py` for zero-orientation testing.

Infrastructure Changes:
- Updated `extension.py` to use a dedicated `xformOp:orient:telemetry` operation, preserving base model rotation.
- Added "Reset Orientation" button to the extension UI.

Bug Fixes:
- Fixed `TypeError` in `lab.py` depth smoothing (float vs int window size).
- Fixed USD error "xformOp already exists" in `extension.py` by checking for existing op.
- Fixed "Incompatible xformOpOrder" warnings by ensuring telemetry rotation is applied before scale.
- Fixed `TypeError` in `extension.py` regarding `addToXformOpOrder`.
- Fixed linting (line length) and type checking (variable reuse) errors in `extension.py`.

UX Improvements:
- Increased default animal model scale from 500 to 5000.

+++ src/biologger_sim/processors/lab.py
@@ -392,7 +392,7 @@ class PostFactoProcessor(BiologgerProcessor):
 
             # Apply 5-second moving average smoothing (R-compatible)
             # R: stats::filter(dat$Depth, filter=rep(1,freq * 5) / (freq * 5))
-            window_size = self.freq * 5
+            window_size = int(self.freq * 5)
             if len(depth_arr) >= window_size:
                 kernel = np.ones(window_size) / window_size
                 # Use mode='same' to match centered filter behavior
